+++ src/main/java/com/montelzek/moneytrack/dto/EditFinancialGoalDTO.java
@@ -7,6 +7,8 @@ import jakarta.validation.constraints.Size;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class EditFinancialGoalDTO {
@@ -17,10 +19,10 @@ public class EditFinancialGoalDTO {
 
     @NotNull(message = "Budget size can't be empty.")
     @DecimalMin(value = "0.01", message = "Amount must be at least 0.01")
-    private Double targetAmount;
+    private BigDecimal targetAmount;
 
     @NotNull(message = "Budget size can't be empty.")
-    private Double currentAmount;
+    private BigDecimal currentAmount;
 
     private Long id;
 }
